Recognise SEC1 PEM files + check PKCS#1 public vs private
SEC1 files have -----EC PRIVATE KEY----- markers and a specific ASN.1
representation that's not currently supported by Ring. At least if
we recognise these markers we can give a meaningful error.

To be consistent with how we cast a PEM data `as_ec_public_key` or
`as_ec_private_key` (and report an error if there is a mismatch
in expectations and the data that's been loaded), this also splits
`as_rsa_key` into `as_rsa_public_key` and `as_rsa_private_key`.