Simplify api by making it synchronous
Originally the idea had been that an Algorithm could be used to abstract
a key set in the future (possibly fetching a jwks url behind the scenes)
and so to allow that the sign/verify APIs were made async.

Since then I've implemented a separate project that handles fetching a
jwks key set for verifying Cognito tokens and it works well enough to
build that on top of a simpler synchronous API.

This also removes the kid: Option<&str> argument for the encode API
which was also only there just in case an Algorithm might abstract a
key set in the future.

An Algorithm can now be associated with a specific kid value (via
alg.set_kid("value")) with the affect that cross checks the kid for
decoded tokens matches or else that's an error.

This helps clarify a more constrained scope for the Algorithm API.