Return Value from Verifier::verify
Instead of automagically deserializing into a custom Claims type the api
now simply returns a serde_json Value. The avoids implicitly,
redundantly deserializing the claims twice when using the recommended
verification api.

It's trivial to use ::from_value() to deserialize into a custom type if
that's _really_ desired - but that also has a notable trade-off in that
it will discard any unknown claims.

This also updates the TokenData type returned by verify_for_time() so
that claims is non-optional. In practice there was no case where this
would ever be None, since the APIs that would be used to just parse
the header don't use this type.