verifier: remove spurious empty string claim name
In the vector of claims that must be of type string there was an
empty string "". I'm not sure what was originally intended here
but it would have been benign to be requiring that a claim with
and empty string name must have a string value. This patch simply
removes this validation.